Correct time calculation bug, and print calculated statistics
Th ts_now variable was being set to 0 on each iteration, so the diff_t
value was wrong, and provoked the yaw to have huge values.

The bias of the gyro is calculated for the selected axis (Z by default)
as well as the standard deviations of the pitch and yaw values.