Add function for getting the blocking of a wave to a flying drone
The function uses basic geometry for getting the distances to a drone
that is not in the horizontal line with the transmitter. It uses then
Fresnel equation for getting the blocking of the wave. It makes the
assumption that the altitude and the fresnel radius are on the same
plane. Minor deviations appear for small angles.